Implement prepared statement
- Add support for prepared statements
- Start supporting for inserts
