--- be/src/main/scala/io/nariok/configuration/Configuration.scala
@@ -1,25 +1,0 @@
-package io.nariok.configuration
-
-import java.io.{File, FileOutputStream}
-
-import com.typesafe.config.{Config, ConfigFactory}
-
-object Configuration {
-  def getConfig(): Config = {
-    val applicationConf = new File("./conf/application.conf")
-    if (applicationConf.exists()) ConfigFactory.parseFile(applicationConf)
-    else {
-      val configBytes = ClassLoader
-        .getSystemResourceAsStream("application.conf")
-        .readAllBytes()
-      val parentDirectory = applicationConf.getParentFile();
-      if (!parentDirectory.exists()) {
-        parentDirectory.mkdir()
-      }
-      applicationConf.createNewFile()
-      val fos = new FileOutputStream(applicationConf)
-      fos.write(configBytes)
-      getConfig()
-    }
-  }
-}
