--- be/src/main/scala/io/nariok/configuration/RawConfiguration.scala
@@ -1,5 +1,0 @@
-package io.nariok.configuration
-
-import com.typesafe.config.Config
-
-final case class RawConfiguration(config: Config)
