--- be/src/main/scala/io/nariok/models/blog/Article.scala
@@ -1,6 +1,0 @@
-package io.nariok.models.blog
-
-final case class Article(id: Int, title: String, contents: List[BlogContent], created: String, modified: String) {
-  def addContent(blogContent: BlogContent): Article =
-    Article(id, title, contents :+ blogContent, created, modified)
-}
