--- be/src/main/scala/io/nariok/models/blog/BlogContent.scala
@@ -1,8 +1,0 @@
-package io.nariok.models.blog
-
-case class BlogContent(id: Int,
-                       value: String,
-                       contentType: String,
-                       blogId: Int,
-                       altText: Option[String] = None,
-                       align: Option[String] = None)
