--- be/src/main/scala/io/nariok/models/blog/BlogContentType.scala
@@ -1,8 +1,0 @@
-package io.nariok.models.blog
-
-sealed trait BlogContentType
-
-case object BlogText extends BlogContentType
-
-case object BlogImage extends BlogContentType
-
