--- be/src/main/scala/io/nariok/models/http/AbstractResponse.scala
@@ -1,3 +1,0 @@
-package io.nariok.models.http
-
-abstract class ApiResponseModel[T](data: Option[T] = None, error: Option[String] = None)
