--- be/src/main/scala/io/nariok/models/http/FailResponse.scala
@@ -1,3 +1,0 @@
-package io.nariok.models.http
-
-case class FailResponse(error: String) extends ApiResponseModel[Nothing](data = None, error = Some(error))
