--- be/src/main/scala/io/nariok/models/http/SuccessResponse.scala
@@ -1,3 +1,0 @@
-package io.nariok.models.http
-
-case class SuccessResponse[T](data: T) extends ApiResponseModel[T](data = Some(data), error = None)
