--- be/src/main/scala/io/nariok/modules/ApplicationModule.scala
@@ -1,21 +1,0 @@
-package io.nariok.modules
-
-import akka.actor.ActorSystem
-import com.google.inject.AbstractModule
-import net.codingwell.scalaguice.ScalaModule
-
-import scala.concurrent.ExecutionContext
-
-class ApplicationModule(implicit executionContext: ExecutionContext, actorSystem: ActorSystem)
-    extends AbstractModule
-    with ScalaModule {
-  override def configure(): Unit = {
-    bind[ActorSystem].toInstance(actorSystem)
-    bind[ExecutionContext].toInstance(executionContext)
-
-    install(new ConfigurationModule())
-    install(new RepositoryModule())
-    install(new ServiceModule())
-    install(new WebRouteModule())
-  }
-}
