--- be/src/main/scala/io/nariok/modules/ServiceModule.scala
@@ -1,12 +1,0 @@
-package io.nariok.modules
-
-import com.google.inject.AbstractModule
-import io.nariok.services.{BlogService, BlogServiceImpl}
-import net.codingwell.scalaguice.ScalaModule
-
-import scala.concurrent.ExecutionContext
-
-class ServiceModule(implicit executionContext: ExecutionContext) extends AbstractModule with ScalaModule {
-  override def configure(): Unit =
-    bind[BlogService].to[BlogServiceImpl].in[com.google.inject.Singleton]
-}
