+++ be/src/main/scala/io/narok/configuration/RawConfiguration.scala
@@ -1,0 +1,5 @@
+package io.narok.configuration
+
+import com.typesafe.config.Config
+
+final case class RawConfiguration(config: Config)
