+++ be/src/main/scala/io/narok/models/blog/Article.scala
@@ -1,0 +1,6 @@
+package io.narok.models.blog
+
+final case class Article(id: Int, title: String, contents: List[BlogContent], created: String, modified: String) {
+  def addContent(blogContent: BlogContent): Article =
+    Article(id, title, contents :+ blogContent, created, modified)
+}
