+++ be/src/main/scala/io/narok/models/blog/BlogContent.scala
@@ -1,0 +1,8 @@
+package io.narok.models.blog
+
+case class BlogContent(id: Int,
+                       value: String,
+                       contentType: String,
+                       blogId: Int,
+                       altText: Option[String] = None,
+                       align: Option[String] = None)
