+++ be/src/main/scala/io/narok/models/blog/BlogContentType.scala
@@ -1,0 +1,7 @@
+package io.narok.models.blog
+
+sealed trait BlogContentType
+
+case object BlogText extends BlogContentType
+
+case object BlogImage extends BlogContentType
