+++ be/src/main/scala/io/narok/models/http/AbstractResponse.scala
@@ -1,0 +1,3 @@
+package io.narok.models.http
+
+abstract class AbstractResponse[T](data: Option[T] = None, error: Option[String] = None)
