+++ be/src/main/scala/io/narok/models/http/FailResponse.scala
@@ -1,0 +1,3 @@
+package io.narok.models.http
+
+case class FailResponse(error: String) extends AbstractResponse[Nothing](data = None, error = Some(error))
