+++ be/src/main/scala/io/narok/models/http/SuccessResponse.scala
@@ -1,0 +1,3 @@
+package io.narok.models.http
+
+case class SuccessResponse[T](data: T) extends AbstractResponse[T](data = Some(data), error = None)
