+++ be/src/main/scala/io/narok/modules/ApplicationModule.scala
@@ -1,0 +1,21 @@
+package io.narok.modules
+
+import akka.actor.ActorSystem
+import com.google.inject.AbstractModule
+import net.codingwell.scalaguice.ScalaModule
+
+import scala.concurrent.ExecutionContext
+
+class ApplicationModule(implicit executionContext: ExecutionContext, actorSystem: ActorSystem)
+    extends AbstractModule
+    with ScalaModule {
+  override def configure(): Unit = {
+    bind[ActorSystem].toInstance(actorSystem)
+    bind[ExecutionContext].toInstance(executionContext)
+
+    install(new ConfigurationModule())
+    install(new RepositoryModule())
+    install(new ServiceModule())
+    install(new WebRouteModule())
+  }
+}
