+++ be/src/main/scala/io/narok/modules/ConfigurationModule.scala
@@ -1,0 +1,13 @@
+package io.narok.modules
+
+import com.google.inject.AbstractModule
+import io.narok.configuration.{Configuration, RawConfiguration}
+import net.codingwell.scalaguice.ScalaModule
+
+class ConfigurationModule extends AbstractModule with ScalaModule {
+  override def configure(): Unit = {
+
+    val rawConfiguration = RawConfiguration(Configuration.getConfig)
+    bind[RawConfiguration].toInstance(rawConfiguration)
+  }
+}
